+++ src/Backend/Modules/Locale/Actions/Add.php
@@ -15,6 +15,7 @@ use Backend\Core\Engine\Model as BackendModel;
 use Backend\Core\Engine\Form as BackendForm;
 use Backend\Core\Engine\Language as BL;
 use Backend\Modules\Locale\Engine\Model as BackendLocaleModel;
+use Common\Uri as CommonUri;
 
 /**
  * This is the add action, it will display a form to add an item to the locale.
@@ -146,7 +147,7 @@ class Add extends BackendBaseActionAdd
             if($txtValue->isFilled(BL::err('FieldIsRequired'))) {
                 // in case this is a 'act' type, there are special rules concerning possible values
                 if($this->frm->getField('type')->getValue() == 'act') {
-                    if(urlencode($txtValue->getValue()) != \CommonUri::getUrl($txtValue->getValue())) $txtValue->addError(BL::err('InvalidValue'));
+                    if(urlencode($txtValue->getValue()) != CommonUri::getUrl($txtValue->getValue())) $txtValue->addError(BL::err('InvalidValue'));
                 }
             }