+++ src/Frontend/Core/Engine/Model.php
@@ -9,6 +9,7 @@ namespace Frontend\Core\Engine;
  * file that was distributed with this source code.
  */
 
+use Common\Cookie as CommonCookie;
 use TijsVerkoyen\Akismet\Akismet;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Filesystem\Exception\IOException;
@@ -659,12 +660,12 @@ class Model extends \BaseModel
         }
 
         // get/init tracking identifier
-        self::$visitorId = \CommonCookie::exists('track') ? (string) \CommonCookie::get('track') : md5(
+        self::$visitorId = CommonCookie::exists('track') ? (string) CommonCookie::get('track') : md5(
             uniqid() . \SpoonSession::getSessionId()
         );
 
-        if (!self::getModuleSetting('Core', 'show_cookie_bar', false) || \CommonCookie::hasAllowedCookies()) {
-            \CommonCookie::set('track', self::$visitorId, 86400 * 365);
+        if (!self::getModuleSetting('Core', 'show_cookie_bar', false) || CommonCookie::hasAllowedCookies()) {
+            CommonCookie::set('track', self::$visitorId, 86400 * 365);
         }
 
         return self::getVisitorId();